[수정] 일일 챌린지를 추가할 때 중복되지않도록, 일일 챌린지를 삭제할때 빈 리스트인지 확인하도록 수정

+++ src/main/java/com/dangdiary/api/domain/schedule/service/impl/ScheduleServiceImp.java
@@ -25,7 +25,9 @@ public class ScheduleServiceImp {
         String date = sdf.format(cal.getTime());
 
         List<Integer> deleteDailyChallengeIds = scheduleDAO.getDeleteUserChallengeIds();
-        scheduleDAO.deleteDailyChallenges(deleteDailyChallengeIds);
+        if (!deleteDailyChallengeIds.isEmpty()) {
+            scheduleDAO.deleteDailyChallenges(deleteDailyChallengeIds);
+        }
         List<Integer> userIds = scheduleDAO.getUserIds();
 
         for (Integer userId: userIds) {
